feat: 정렬 로직 분리
